+++ fast-company/src/components/users.jsx
@@ -2,14 +2,45 @@ import React, { useState, useEffect } from "react"
 import Pagination from "./pagination"
 import SearchStatus from "./searchstatus"
 import UsersTables from "./usersTable"
+import GroupList from "./grouplist"
+import Loader from "./loader"
 import { paginate } from "../utils/paginate"
-import PropTypes from "prop-types"
+import API from "../api"
 import _ from "lodash"
 
-const Users = ({ bookmarkActive, deleteUser, users, filteredItems }) => {
+const Users = () => {
     const [currentPage, setCurrentPage] = useState(1)
+    const [users, setUsers] = useState([])
+    const [professions, setProfessions] = useState()
+    const [filteredItems, setFilteredItems] = useState()
     const [order, setOrder] = useState({ path: "", order: "asc" })
     const pageSize = 6
+    const handleItems = (selected) => {
+        setFilteredItems(selected)
+    }
+
+    const handleDelete = (userId) => {
+        setUsers(users.filter((user) => user._id !== userId))
+    }
+
+    const handleReset = () => {
+        setFilteredItems()
+    }
+
+    const handleToggleBookmark = (userId) => {
+        setUsers((prevState) =>
+            prevState.map((user) => {
+                if (user._id === userId) {
+                    return {
+                        ...user,
+                        bookmark: !user.bookmark
+                    }
+                } else {
+                    return user
+                }
+            })
+        )
+    }
     const handlePageChange = (page) => {
         setCurrentPage(page)
     }
@@ -22,6 +53,17 @@ const Users = ({ bookmarkActive, deleteUser, users, filteredItems }) => {
     const sorteredUsers = _.orderBy(filtered, order.path, order.order)
     const userCrop = paginate(sorteredUsers, currentPage, pageSize)
     const count = filtered.length
+
+    useEffect(() => {
+        API.users.fetchAll().then((data) => {
+            setUsers(data)
+        })
+    }, [])
+    useEffect(() => {
+        API.professions.fetchAll().then((data) => {
+            setProfessions(data)
+        })
+    }, [])
     useEffect(() => {
         if (userCrop.length === 0 && currentPage !== 1) {
             setCurrentPage(currentPage - 1)
@@ -31,26 +73,40 @@ const Users = ({ bookmarkActive, deleteUser, users, filteredItems }) => {
     return (
         <>
             <SearchStatus length={count} />
-            <UsersTables
-                users={userCrop}
-                bookmarkActive={bookmarkActive}
-                deleteUser={deleteUser}
-                onSort={handleSort}
-                selectedSort={order}
-            />
-            <Pagination
-                itemsCount={count}
-                pageSize={pageSize}
-                onPageChange={handlePageChange}
-                active={currentPage}
-            />
+            <div className="d-flex mt-2">
+                {professions ? (
+                    <div className="mx-2">
+                        <GroupList
+                            items={professions}
+                            checkItems={handleItems}
+                            reset={handleReset}
+                            active={filteredItems}
+                        />
+                    </div>
+                ) : (
+                    <div>
+                        <Loader string={"загружаем профессии"} />
+                    </div>
+                )}
+
+                <div className="d-flex flex-column flex-grow-1 m-2">
+                    <UsersTables
+                        users={userCrop}
+                        bookmarkActive={handleToggleBookmark}
+                        deleteUser={handleDelete}
+                        onSort={handleSort}
+                        selectedSort={order}
+                    />
+                    <Pagination
+                        itemsCount={count}
+                        pageSize={pageSize}
+                        onPageChange={handlePageChange}
+                        active={currentPage}
+                    />
+                </div>
+            </div>
         </>
     )
 }
-Users.propTypes = {
-    bookmarkActive: PropTypes.func.isRequired,
-    deleteUser: PropTypes.func.isRequired,
-    users: PropTypes.array.isRequired,
-    filteredItems: PropTypes.object
-}
+
 export default Users
